# file-api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: File API
  description: Test API for retrieving files and file counts from local folders. This API is used for documentation in jeremiah-michaels/resume.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /files/{folder}/{filename}:
    get:
      summary: Get a file
      description: Retrieve a file by name from a specified folder inside the files directory.
      parameters:
        - name: folder
          in: path
          required: true
          description: The name of the folder to look in.
          schema:
            type: string
            example: folder1
        - name: filename
          in: path
          required: true
          description: The name of the file to retrieve.
          schema:
            type: string
            example: doc1.txt
      responses:
        "200":
          description: The requested file.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /files/{folder}/count:
    get:
      summary: Get file count
      description: Returns the number of files in a specified folder inside the files directory.
      parameters:
        - name: folder
          in: path
          required: true
          description: The name of the folder to count files in.
          schema:
            type: string
            example: folder1
      responses:
        "200":
          description: The file count for the specified folder.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileCount"
        "400":
          description: Invalid path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Folder not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    FileCount:
      type: object
      properties:
        folder:
          type: string
          example: folder1
        file_count:
          type: integer
          example: 3

    Error:
      type: object
      properties:
        detail:
          type: string
          example: File 'doc1.txt' not found in folder 'folder1'.